wayland: move ExtImageCopyBackend class and related types to the header
